initial rework
