+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "vercel-rust",
   "description": "Rust runtime for Vercel Functions.",
-  "version": "3.1.2",
+  "version": "4.0.0",
   "license": "MIT",
   "main": "./dist/index",
   "homepage": "https://github.com/vercel-community/rust",
@@ -10,7 +10,7 @@
     "url": "https://github.com/vercel-community/rust.git"
   },
   "scripts": {
-    "build": "tsc",
+    "build": "tsc --removeComments",
     "prepublishOnly": "tsc",
     "test": "jest",
     "typecheck": "tsc --noEmit"
@@ -20,26 +20,13 @@
     "launcher.rs"
   ],
   "dependencies": {
-    "@iarna/toml": "^2.2.5",
-    "execa": "^5.0.0",
-    "fs-extra": "^7.0.1"
+    "execa": "5",
+    "typescript": "^4.9.4"
   },
   "devDependencies": {
-    "@types/fs-extra": "^7.0.0",
-    "@vercel/build-utils": "^2.6.0",
-    "husky": "^3.1.0",
-    "jest": "^25.5.4",
-    "lint-staged": "^9.5.0",
-    "ms": "^2.1.2",
-    "node-fetch": "^2.6.0",
-    "prettier": "^2.3.2",
-    "typescript": "3.5.2",
-    "vercel": "^23.1.2"
-  },
-  "prettier": {
-    "arrowParens": "always",
-    "singleQuote": true,
-    "tabWidth": 2,
-    "trailingComma": "none"
+    "@types/node": "^18.11.18",
+    "@vercel/build-utils": "^6.2.1",
+    "@vercel/routing-utils": "^2.1.8",
+    "rimraf": "^4.1.1"
   }
 }
