--- package/CHANGELOG.md
@@ -1,9 +1,0 @@
-# vercel-rust changelog
-
-## Unreleased
-
-## 0.2.0
-
-- Renamed to vercel_lambda
-
-## All other versions are lost to time. Sorry!
