--- package/launcher.rs
@@ -1,8 +1,0 @@
-use vercel_lambda::lambda;
-use std::error::Error;
-
-// PLACEHOLDER
-
-fn main() -> Result<(), Box<dyn Error>> {
-    Ok(lambda!(handler))
-}
