+++ src/lib/runtime.ts
@@ -1,0 +1,19 @@
+import {
+  BuildOptions,
+  BuildResultV3,
+  Files,
+  PrepareCacheOptions,
+  ShouldServeOptions,
+  StartDevServerOptions,
+  StartDevServerResult,
+} from "@vercel/build-utils/dist";
+
+export interface Runtime {
+  version: number;
+  build: (options: BuildOptions) => Promise<BuildResultV3>;
+  prepareCache?: (options: PrepareCacheOptions) => Promise<Files>;
+  shouldServe?: (options: ShouldServeOptions) => Promise<boolean>;
+  startDevServer?: (
+    options: StartDevServerOptions
+  ) => Promise<StartDevServerResult>;
+}
