--- tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": false,
-    "esModuleInterop": true,
-    "lib": ["esnext"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noEmitOnError": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitReturns": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "outDir": "dist",
-    "sourceMap": false,
-    "strict": true,
-    "target": "esnext",
-    "types": ["node"],
-    "skipLibCheck": true
-  }
-}
