+++ vercel_runtime/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "vercel_runtime"
+version = "0.1.2"
+edition = "2021"
+description = "Vercel Function Runtime"
+license = "MIT"
+exclude = [
+  "tests/*",
+]
+
+[dependencies]
+lambda_http = { version = "0.7", default-features = false, features = ["apigw_http"] }
+lambda_runtime = "0.7"
+tokio = { version = "1", features = ["macros"] }
+tracing = { version = "0.1", features = ["log"] }
+tracing-subscriber = { version = "0.3", default-features = false, features = ["fmt"] }
+serde = { version = "1.0.145", features = ["derive"] }
+serde_json = { version = "1.0.86", features = ["raw_value"] }
+serde_derive = "1.0.9"
+tower-http = { version = "0.3.4", features = ["cors"] }
+tower-service = "0.3.2"
+base64 = "0.10"
+bytes = "1.3.0"
+async-trait = "0.1.63"
+
